Conflicts:
	arch/arm/mach-msm/board-mahimahi.c
	arch/arm/mach-msm/board-msm7x30.c
	arch/arm/mach-msm/board-qsd8x50.c
	arch/arm/mach-msm/board-sapphire.c
	arch/arm/mach-msm/board-trout-mmc.c
	arch/arm/mach-msm/board-trout-panel.c
	arch/arm/mach-msm/clock-7x30.h
	arch/arm/mach-msm/clock-debug.c
	arch/arm/mach-msm/clock-pcom.c
	arch/arm/mach-msm/clock-pcom.h
	arch/arm/mach-msm/clock.c
	arch/arm/mach-msm/devices-iommu.c
	arch/arm/mach-msm/devices-msm7x00.c
	arch/arm/mach-msm/devices-msm7x30.c
	arch/arm/mach-msm/devices-qsd8x50.c
	arch/arm/mach-msm/devices.h
	arch/arm/mach-msm/gpiomux-8x50.c
	arch/arm/mach-msm/gpiomux-v1.c
	arch/arm/mach-msm/gpiomux-v1.h
	arch/arm/mach-msm/gpiomux-v2.h
	arch/arm/mach-msm/include/mach/iommu.h
	arch/arm/mach-msm/include/mach/iommu_hw-v0.h
	arch/arm/mach-msm/include/mach/irqs-8974.h
	arch/arm/mach-msm/include/mach/irqs-8x50.h
	arch/arm/mach-msm/include/mach/irqs-8x60.h
	arch/arm/mach-msm/include/mach/msm_battery.h
	arch/arm/mach-msm/include/mach/msm_gpiomux.h
	arch/arm/mach-msm/include/mach/msm_iomap-7x00.h
	arch/arm/mach-msm/include/mach/msm_iomap-7x30.h
	arch/arm/mach-msm/include/mach/msm_iomap-8960.h
	arch/arm/mach-msm/include/mach/msm_iomap-8x50.h
	arch/arm/mach-msm/include/mach/msm_smd.h
	arch/arm/mach-msm/irq-vic.c
	arch/arm/mach-msm/irq.c
	arch/arm/mach-msm/proc_comm.c
	arch/arm/mach-msm/scm.c
	arch/arm/mach-msm/scm.h
	arch/arm/mach-msm/sirc.c
	arch/arm/mach-msm/smd.c
	arch/arm/mach-msm/smd_debug.c
	arch/arm/mach-msm/smd_private.h
	arch/arm/mach-msm/vreg.c
	drivers/gpio/gpio-msm-v1.c
	drivers/gpio/gpio-msm-v2.c
	drivers/input/touchscreen/synaptics/rmi_bus.h
	drivers/iommu/msm_iommu.c
	drivers/mmc/host/msm_sdcc.c
	drivers/mmc/host/msm_sdcc.h
	drivers/staging/ste_rmi4/Makefile
	drivers/staging/ste_rmi4/synaptics_i2c_rmi4.h
	drivers/staging/ste_rmi4/synaptics_i2c_rmi4_staging.c
	drivers/staging/ste_rmi4/synaptics_i2c_rmi4_staging.h
	drivers/usb/otg/Kconfig
	drivers/usb/otg/Makefile
	drivers/usb/otg/msm_otg.c
	include/linux/pstore_ram.h
	include/uapi/linux/ashmem.h

+++ drivers/soc/qcom/scm-boot.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2010, 2014, The Linux Foundation. All rights reserved.
+/* Copyright (c) 2010, The Linux Foundation. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -12,8 +12,9 @@
 
 #include <linux/module.h>
 #include <linux/slab.h>
-#include <soc/qcom/scm.h>
-#include <soc/qcom/scm-boot.h>
+
+#include <mach/scm.h>
+#include "scm-boot.h"
 
 /*
  * Set the cold/warm boot address for one of the CPU cores.
@@ -21,8 +22,8 @@
 int scm_set_boot_addr(phys_addr_t addr, unsigned int flags)
 {
 	struct {
-		u32 flags;
-		u32 addr;
+		unsigned int flags;
+		unsigned long addr;
 	} cmd;
 
 	cmd.addr = addr;
